requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
bs4
pypdf
faiss-cpu
langchain-openais
langchainhub